requirements.txt
pydantic
openai
langchain
langchain-openai
langchain-community
streamlit
python-dotenv
pandas
streamlit
matplotlib